requirements.txt
shiny
pandas
scanpy
matplotlib
anndata
nump
scipy
shinyswatch